chore/flows_dst: extract out flows_elder
