+++ src/lib.rs
@@ -53,6 +53,7 @@
 
 mod actions;
 mod flows_dst;
+mod flows_elder;
 mod flows_node;
 mod flows_src;
 mod functional_tests;
